feat: 핵심 로직, UI 로직 클래스 분리
