+++ src/main/java/racing/InputView.java
@@ -1,0 +1,4 @@
+package racing;
+
+public class InputView {
+}
