+++ src/main/java/racing/RacingGame.java
@@ -1,33 +1,14 @@
 package racing;
 
 import java.util.Random;
-import java.util.Scanner;
+
+import static racing.RacingMain.GO_CONDITION;
+import static racing.RacingMain.RANDOM_MAX;
 
 public class RacingGame {
-    public static void main(String[] args) {
-        System.out.println("자동차 대수는 몇 대 인가요?");
-        Scanner scanner = new Scanner(System.in);
-        int carCount = scanner.nextInt();
-        System.out.println("시도할 횟수는 몇 회인가요?");
-        int tryCount = scanner.nextInt();
-
-        System.out.println();
-        System.out.println("실행 결과");
-
-        String[] carPosition = new String[carCount];
-        for (int i = 0; i < carCount; i++) {
-            carPosition[i] = "";
-        }
-
-        for (int i = 0; i < tryCount; i++) {
-            for (int j = 0; j < carCount; j++) {
-                if (new Random().nextInt(10) >= 4) {
-                    carPosition[j] += '-';
-                }
-                System.out.println(carPosition[j]);
-            }
-//            System.out.println(Arrays.toString(carPosition));
-            System.out.println();
+    static void carMove(String[] carPosition, int j) {
+        if (new Random().nextInt(RANDOM_MAX) >= GO_CONDITION) {
+            carPosition[j] += '-';
         }
     }
 }
