+++ src/test/java/racing/RacingGameTest.java
@@ -2,9 +2,9 @@ package racing;
 
 import org.junit.jupiter.api.Test;
 
-public class RacingGameTest {
+class RacingGameTest {
     @Test
     void carInput() {
-        int result = RacingGame.carInput();
+        int result = RacingGame.carMove();
     }
 }
